fix: Code View Command
